update: 初期設定
