+++ lib/main.dart
@@ -1,5 +1,5 @@
 import 'package:flutter/material.dart';
-import 'package:randomspotfinder/screens/initial_screen.dart';
+import 'package:randomspotfinder/presentation/pages/login/initial_screen.dart';
 
 void main() {
   runApp(const MyApp());
