+++ lib/presentation/pages/home/home_screen.dart
@@ -1,0 +1,11 @@
+
+import 'package:flutter/cupertino.dart';
+
+class Home extends StatelessWidget {
+  const Home({super.key});
+
+  @override
+  Widget build(BuildContext context) {
+    return const Placeholder();
+  }
+}
