+++ lib/presentation/pages/location_setting/search_radius_setting_screen.dart
@@ -1,7 +1,7 @@
 //TODO: 検索半径設定画面
 import 'package:flutter/cupertino.dart';
 import 'package:flutter/material.dart';
-import 'package:randomspotfinder/screens/roulette_screen.dart';
+import 'package:randomspotfinder/presentation/pages/roulette/roulette_screen.dart';
 
 class SearchRadiusSettingScreen extends StatelessWidget {
   @override
