+++ pubspec.yaml
@@ -12,6 +12,7 @@ dependencies:
 
   cupertino_icons: ^1.0.2
   firebase_core: ^2.27.0
+  cloud_firestore: ^4.16.0
 
 dev_dependencies:
   flutter_test:
